Clean up coverage and fix issue with notebook names
