+++ extensions/notebook/src/common/constants.ts
@@ -59,3 +59,5 @@ export enum PythonPkgType {
 export const pythonWindowsInstallUrl = 'https://go.microsoft.com/fwlink/?linkid=2110625';
 export const pythonMacInstallUrl = 'https://go.microsoft.com/fwlink/?linkid=2128152';
 export const pythonLinuxInstallUrl = 'https://go.microsoft.com/fwlink/?linkid=2110524';
+
+export const notebookLanguages = ['notebook', 'ipynb'];
